requirements.txt
fastapi
uvicorn
langchain
openai
pydantic
python-dotenv
langchain-openai
